corrreções

+++ Src/api-application-labmark/Domain/Modules/Exam/Infrastructure/Models/Dtos/ColiformsEscherichiaDto.cs
@@ -9,11 +9,11 @@ namespace Labmark.Domain.Modules.Exam.Infrastructure.Models.Dtos
         public int? Point { get; set; }
         public double? Brilla { get; set; }
         public int BOD { get; set; }
-        public int? WatherBath { get; set; }
+        public double? WatherBath { get; set; }
         public int? FlowMicropipettor { get; set; }
         public double? Escherichia { get; set; }
-        public int? TotalColifoms { get; set; }
-        public int? TolerantColiforms { get; set; }
+        public double? TotalColifoms { get; set; }
+        public double? TolerantColiforms { get; set; }
         public DateTime? DateOfCompletion { get; set; }
         public double? Result { get; set; }
         public DateTime? DateResult { get; set; }
